Refactor FilmService; add RedisService & ElasticService.
