+++ src/db/elastic.py
@@ -3,20 +3,25 @@ from abc import ABC, abstractmethod
 from elasticsearch import AsyncElasticsearch
 
 
-class AsyncSearchInterface(ABC):
-    @abstractmethod
-    async def get_search_connection(self):
+class AsyncSearchAbstract(ABC):
+ #   @abstractmethod
+  #  async def get_search_connection(self):
         pass
 
     
-class AsyncSearchImplementation(AsyncSearchInterface):
-    async def get_search_connection(self):
-        return AsyncElasticsearch()
+#class AsyncSearchImplementation(AsyncSearchInterface):
+#    def __init__(self, elastic):
+#        self.elastic = elastic
 
 
-async def main():
-    search_implementation = AsyncSearchImplementation()
-    es = search_implementation.get_search_connection()
-    return es
+elastic: AsyncElasticsearch = AsyncElasticsearch(hosts='http://localhost:9200/')
 
-es = AsyncSearchImplementation()
+async def get_elastic() -> AsyncElasticsearch:
+    return elastic
+
+#async def main():
+#    search_implementation = AsyncSearchImplementation()
+#    es = search_implementation.get_search_connection()
+#    return es
+
+#es = AsyncSearchImplementation()
