+++ src/db/redis.py
@@ -1,22 +1,31 @@
-from abc import ABC, abstractmethod
+from abc import ABC
 
 from redis.asyncio import Redis
 
 
-class AsyncCacheInterface(ABC):
-    @abstractmethod
-    async def get_cache_connection(self):
+class AsyncCacheAbstract(ABC):
+#    @abstractmethod
+#    async def get_cache_connection(self):
         pass
 
 
-class AsyncCacheImplementation(AsyncCacheInterface):
-    async def get_cache_connection(self):
-        return Redis()
+#class AsyncCacheImplementation(AsyncCacheInterface):
+#    def __init__(self, redis):
+#        self.redis = redis
+
+#    async def get_cache_connection(self):
+#        return Redis()
 
 
-async def main():
-    cache_implementation = AsyncCacheImplementation()
-    conn = await cache_implementation.get_cache_connection()
-    return conn
+redis: Redis = Redis()
 
-redis = AsyncCacheImplementation()
+async def get_redis() -> Redis:
+    return redis
+
+
+#async def main():
+#    cache_implementation = AsyncCacheImplementation()
+#    conn = await cache_implementation.get_cache_connection()
+#    return conn
+
+#redis = AsyncCacheImplementation()
