+++ src/main.py
@@ -21,8 +21,8 @@ app = FastAPI(
 
 @app.on_event('startup')
 async def startup():
-    redis.redis = redis.AsyncCacheImplementation(host=settings.REDIS_HOST, port=settings.REDIS_PORT)
-    elastic.es = elastic.AsyncSearchImplementation(
+    redis.rdedis = Redis(host=settings.REDIS_HOST, port=settings.REDIS_PORT)
+    elastic.es = AsyncElasticsearch(
         hosts=[
             f'{settings.ELASTIC_SCHEME}://'
             f'{settings.ELASTIC_HOST}'
@@ -40,8 +40,8 @@ async def shutdown():
 # Подключаем роутер к серверу, указав префикс /v1/films
 # Теги указываем для удобства навигации по документации
 app.include_router(films.router, prefix='/api/v1/films', tags=['films'])
-app.include_router(genres.router, prefix='/api/v1/genres', tags=['genres'])
-app.include_router(persons.router, prefix='/api/v1/persons', tags=['persons'])
+#app.include_router(genres.router, prefix='/api/v1/genres', tags=['genres'])
+#app.include_router(persons.router, prefix='/api/v1/persons', tags=['persons'])
 
 if __name__ == '__main__':
     uvicorn.run(
